Move code to BaseChatController. Add support for all channel domains in getChatChannelSelectionService
