+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/bisq_easy/private_chats/BisqEasyPrivateChatsController.java
@@ -31,11 +31,6 @@ public final class BisqEasyPrivateChatsController extends PrivateChatsController
     }
 
     @Override
-    public void createDependencies(ChatChannelDomain chatChannelDomain) {
-        selectionService = chatService.getBisqEasyPrivateChatChannelSelectionService();
-    }
-
-    @Override
     public BisqEasyPrivateChatsModel createAndGetModel(ChatChannelDomain chatChannelDomain) {
         return new BisqEasyPrivateChatsModel(chatChannelDomain);
     }
