+++ apps/desktop/desktop/src/main/java/bisq/desktop/main/content/chat/BaseChatController.java
@@ -18,11 +18,9 @@
 package bisq.desktop.main.content.chat;
 
 import bisq.bisq_easy.NavigationTarget;
-import bisq.chat.ChatChannel;
-import bisq.chat.ChatChannelDomain;
-import bisq.chat.ChatMessage;
-import bisq.chat.ChatService;
+import bisq.chat.*;
 import bisq.chat.common.CommonPublicChatChannel;
+import bisq.common.observable.Pin;
 import bisq.desktop.ServiceProvider;
 import bisq.desktop.common.threading.UIThread;
 import bisq.desktop.common.view.Navigation;
@@ -36,6 +34,7 @@ import bisq.user.profile.UserProfile;
 import bisq.user.profile.UserProfileService;
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
+import org.fxmisc.easybind.EasyBind;
 import org.fxmisc.easybind.Subscription;
 
 import javax.annotation.Nullable;
@@ -50,11 +49,13 @@ public abstract class BaseChatController<V extends BaseChatView, M extends BaseC
     protected final V view;
     protected final ServiceProvider serviceProvider;
     protected final ChatService chatService;
+    protected final ChatChannelSelectionService selectionService;
     protected final UserIdentityService userIdentityService;
     protected final UserProfileService userProfileService;
     protected final ChannelSidebar channelSidebar;
     protected final ChatMessageContainerController chatMessageContainerController;
     protected Subscription searchTextPin;
+    protected Pin selectedChannelPin;
 
     public BaseChatController(ServiceProvider serviceProvider,
                               ChatChannelDomain chatChannelDomain,
@@ -63,6 +64,7 @@ public abstract class BaseChatController<V extends BaseChatView, M extends BaseC
 
         this.serviceProvider = serviceProvider;
         chatService = serviceProvider.getChatService();
+        selectionService = chatService.getChatChannelSelectionService(chatChannelDomain);
         userIdentityService = serviceProvider.getUserService().getUserIdentityService();
         userProfileService = serviceProvider.getUserService().getUserProfileService();
 
@@ -83,7 +85,27 @@ public abstract class BaseChatController<V extends BaseChatView, M extends BaseC
         view = createAndGetView();
     }
 
-    public abstract void createDependencies(ChatChannelDomain chatChannelDomain);
+    @Override
+    public void onActivate() {
+        selectedChannelPin = selectionService.getSelectedChannel().addObserver(this::selectedChannelChanged);
+        model.getSearchText().set("");
+        searchTextPin = EasyBind.subscribe(model.getSearchText(), searchText -> {
+            if (searchText == null || searchText.isEmpty()) {
+                chatMessageContainerController.setSearchPredicate(item -> true);
+            } else {
+                chatMessageContainerController.setSearchPredicate(item -> item.match(searchText));
+            }
+        });
+    }
+
+    @Override
+    public void onDeactivate() {
+        selectedChannelPin.unbind();
+        searchTextPin.unsubscribe();
+    }
+
+    public void createDependencies(ChatChannelDomain chatChannelDomain) {
+    }
 
     public abstract M createAndGetModel(ChatChannelDomain chatChannelDomain);
